+++ src/components/PupilViews/PupilReview/PupilReview.view.tsx
@@ -1,3 +1,4 @@
+import { useState } from "react";
 import {
   OakFlex,
   OakGrid,
@@ -12,7 +13,10 @@ import {
   OakPrimaryButton,
   OakTertiaryButton,
 } from "@oaknational/oak-components";
-import { useOakPupil } from "@oaknational/oak-pupil-client";
+import {
+  attemptDataCamelCaseSchema,
+  useOakPupil,
+} from "@oaknational/oak-pupil-client";
 import { useFeatureFlagEnabled } from "posthog-js/react";
 
 import { PupilExperienceViewProps } from "../PupilExperience";
@@ -25,8 +29,6 @@ import { QuestionsArray } from "@/components/PupilComponents/QuizEngineProvider"
 import { QuizResults } from "@/components/PupilComponents/QuizResults";
 import { resolveOakHref } from "@/common-lib/urls";
 import { CopyrightNotice } from "@/components/PupilComponents/CopyrightNotice";
-
-// TODO: add question arrays for starter and exit quizzes so that the expand quiz results can be rendered
 
 type PupilViewsReviewProps = {
   lessonTitle: string;
@@ -45,10 +47,7 @@ export const PupilViewsReview = (props: PupilViewsReviewProps) => {
     backUrl,
     starterQuizQuestionsArray,
     exitQuizQuestionsArray,
-    programmeSlug,
-    unitSlug,
     browseData: { programmeFields, lessonSlug, isLegacy },
-    pageType,
   } = props;
   const { phase = "primary", yearDescription, subject } = programmeFields;
   const {
@@ -67,6 +66,7 @@ export const PupilViewsReview = (props: PupilViewsReviewProps) => {
   const pupilClient = useOakPupil();
   const { logAttempt } = pupilClient;
   const isShowShareButtons = useFeatureFlagEnabled("share-results-button");
+  const [isAttemptingShare, setIsAttemptingShare] = useState<boolean>(false);
 
   const bottomNavSlot = (
     <OakLessonBottomNav>
@@ -82,27 +82,35 @@ export const PupilViewsReview = (props: PupilViewsReviewProps) => {
     </OakLessonBottomNav>
   );
   const handleShareResultsClick = async () => {
+    setIsAttemptingShare(true);
     const attemptData = {
       lessonData: { slug: lessonSlug, title: lessonTitle },
       browseData: { subject: subject, yearDescription: yearDescription ?? "" },
       sectionResults: sectionResults,
     };
     try {
-      const attemptId = await logAttempt(attemptData, false);
+      const parsedAttemptData = attemptDataCamelCaseSchema.parse(attemptData);
+      const attemptId = await logAttempt(parsedAttemptData, false);
       if (!attemptId) {
         throw new Error("Failed to log attempt");
       }
+      setIsAttemptingShare(false);
       const shareUrl = `${
         process.env.NEXT_PUBLIC_CLIENT_APP_BASE_URL
       }${resolveOakHref({
-        page: "pupil-lesson-results-canonical",
+        page: "pupil-lesson-results-canonical-share",
         lessonSlug,
         attemptId,
       })}`;
-      alert("See results at " + shareUrl);
+      setTimeout(() => {
+        alert("See results at " + shareUrl);
+      }, 0);
     } catch (e) {
+      setIsAttemptingShare(false);
       console.error(e);
-      alert("Failed to share results");
+      setTimeout(() => {
+        alert("Failed to log attempt");
+      }, 0);
     }
   };
 
@@ -112,22 +120,15 @@ export const PupilViewsReview = (props: PupilViewsReviewProps) => {
       browseData: { subject: subject, yearDescription: yearDescription ?? "" },
       sectionResults: sectionResults,
     };
-    const attemptId = await logAttempt(attemptData, true);
+    const parsedAttemptData = attemptDataCamelCaseSchema.parse(attemptData);
+    const attemptId = await logAttempt(parsedAttemptData, true);
     if (attemptId)
       window.open(
-        pageType === "canonical"
-          ? resolveOakHref({
-              page: "pupil-lesson-results-canonical",
-              lessonSlug,
-              attemptId,
-            })
-          : resolveOakHref({
-              page: "pupil-lesson-results",
-              programmeSlug,
-              unitSlug,
-              lessonSlug,
-              attemptId,
-            }),
+        resolveOakHref({
+          page: "pupil-lesson-results-canonical-printable",
+          lessonSlug,
+          attemptId,
+        }),
         "_blank",
       );
   };
@@ -190,6 +191,7 @@ export const PupilViewsReview = (props: PupilViewsReviewProps) => {
                     title="Share results"
                     onClick={handleShareResultsClick}
                     data-testid="share-results-button"
+                    isLoading={isAttemptingShare}
                   >
                     Share results
                   </OakPrimaryButton>