+++ src/pages/pupils/lessons/[lessonSlug]/results/[attemptId]/printable.tsx
@@ -1,0 +1,5 @@
+const CanonicalPrintableResultsPage = () => {
+  return <div>CanonicalPrintableResultsPage</div>;
+};
+
+export default CanonicalPrintableResultsPage;